theravoice-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic>=2.6.1,<3.0.0
pydantic[email]>=2.6.1,<3.0.0
supabase==1.2.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
requests==2.31.0
pymongo==4.6.1
bcrypt==4.1.2
elevenlabs>=1.0.0,<2.0.0
python-multipart==0.0.9 